Refactor some code in integration tests
