+++ packages/framework-integration-tests/integration/helper/sleep.ts
@@ -1,3 +1,31 @@
 export async function sleep(ms: number): Promise<void> {
   return new Promise((resolve) => setTimeout(resolve, ms))
 }
+
+export async function waitForIt<TResult>(
+  tryFunction: () => Promise<TResult>,
+  checkResult: (result: TResult) => boolean,
+  trialDelayMs = 1000,
+  timeoutMs = 60000
+): Promise<TResult> {
+  console.debug('[waitForIt] start')
+  const start = Date.now()
+  return doWaitFor()
+
+  async function doWaitFor(): Promise<TResult> {
+    console.debug('.')
+    const res = await tryFunction()
+    if (checkResult(res)) {
+      console.debug('[waitForIt] match!')
+      return res
+    }
+    const elapsed = Date.now() - start
+
+    if (elapsed > timeoutMs) {
+      throw new Error('[waitForIt] Timeout reached')
+    }
+
+    await sleep(trialDelayMs)
+    return doWaitFor()
+  }
+}
