+++ packages/framework-integration-tests/integration/providers/aws/functionality/subscriptions.integration.ts
@@ -3,13 +3,14 @@ import {
   countSubscriptionsItems,
   DisconnectableApolloClient,
   graphQLClientWithSubscriptions,
-  waitForIt,
+
 } from '../utils'
 import { random } from 'faker'
 import gql from 'graphql-tag'
 import { expect } from 'chai'
 import * as chai from 'chai'
 import { Observable } from 'apollo-client/util/Observable'
+import { waitForIt } from '../../../helper/sleep'
 
 chai.use(require('chai-as-promised'))
