+++ packages/framework-integration-tests/integration/providers/aws/utils.ts
@@ -1,21 +1,21 @@
-import { CloudFormation, CognitoIdentityServiceProvider, DynamoDB, config } from 'aws-sdk'
+import { CloudFormation, CognitoIdentityServiceProvider, config, DynamoDB } from 'aws-sdk'
 import { Stack, StackResourceDetail, StackResourceSummary } from 'aws-sdk/clients/cloudformation'
 import { ApolloClient } from 'apollo-client'
 import { InMemoryCache, NormalizedCacheObject } from 'apollo-cache-inmemory'
 import { HttpLink } from 'apollo-link-http'
 import fetch from 'cross-fetch'
 import { DocumentClient } from 'aws-sdk/lib/dynamodb/document_client'
-import ScanOutput = DocumentClient.ScanOutput
-import QueryOutput = DocumentClient.QueryOutput
 import { internet } from 'faker'
-import { sleep } from '../../helper/sleep'
 import { WebSocketLink } from 'apollo-link-ws'
 import { getMainDefinition } from 'apollo-utilities'
-import { split, ApolloLink } from 'apollo-link'
+import { ApolloLink, split } from 'apollo-link'
 import * as WebSocket from 'ws'
 import { SubscriptionClient } from 'subscriptions-transport-ws'
 import { ApolloClientOptions } from 'apollo-client/ApolloClient'
 import util = require('util')
+import QueryOutput = DocumentClient.QueryOutput
+import ScanOutput = DocumentClient.ScanOutput
+
 const exec = util.promisify(require('child_process').exec)
 
 const cloudFormation = new CloudFormation()
@@ -569,30 +569,3 @@ export async function getEventsByEntityId(entityID: string): Promise<any> {
 
 // --- Other helpers ---
 
-export async function waitForIt<TResult>(
-  tryFunction: () => Promise<TResult>,
-  checkResult: (result: TResult) => boolean,
-  trialDelayMs = 1000,
-  timeoutMs = 60000
-): Promise<TResult> {
-  console.debug('[waitForIt] start')
-  const start = Date.now()
-  return doWaitFor()
-
-  async function doWaitFor(): Promise<TResult> {
-    console.debug('.')
-    const res = await tryFunction()
-    if (checkResult(res)) {
-      console.debug('[waitForIt] match!')
-      return res
-    }
-    const elapsed = Date.now() - start
-
-    if (elapsed > timeoutMs) {
-      throw new Error('[waitForIt] Timeout reached')
-    }
-
-    await sleep(trialDelayMs)
-    return doWaitFor()
-  }
-}
