+++ packages/framework-integration-tests/integration/providers/local/read-models.integration.ts
@@ -4,9 +4,9 @@ import { changeCartItem, graphQLClient } from './utils'
 import { random } from 'faker'
 import { expect } from 'chai'
 import * as DataStore from 'nedb'
-import { waitForIt } from '../aws/utils'
 import { sandboxPath } from './constants'
 import * as path from 'path'
+import { waitForIt } from '../../helper/sleep'
 
 describe('read-models', () => {
   let readModels: DataStore<unknown>
